Fixed EKF. Added eigen-ros conversions from Willow Garage

+++ auv_core/include/auv_core/rot3d.hpp
@@ -18,11 +18,11 @@ Eigen::Quaterniond rpy2Quat(double roll, double pitch, double yaw);
 
 Eigen::Vector3d quat2RPY(const Eigen::Quaterniond &quaternion);
 
-Eigen::Matrix3f getRotationMat(int axis, float angle);
+Eigen::Matrix3d getRotationMat(int axis, double angle);
 
-Eigen::Matrix3f getEulerRotationMat(const Eigen::Ref<const Eigen::Vector3f> &rpy);
+Eigen::Matrix3d getEulerRotationMat(const Eigen::Ref<const Eigen::Vector3d> &rpy);
 
-Eigen::Matrix3f skewSym(const Eigen::Ref<const Eigen::Vector3f> &v);
+Eigen::Matrix3d skewSym(const Eigen::Ref<const Eigen::Vector3d> &v);
 
 float sawtoothWave(float x, float period, float max);
 
@@ -32,11 +32,11 @@ float mapRollYaw(float x);
 
 float mapPitch(float x);
 
-Eigen::Vector3f getConstrainedRPY(const Eigen::Ref<const Eigen::Vector3f> attitude);
+Eigen::Vector3d getConstrainedRPY(const Eigen::Ref<const Eigen::Vector3d> attitude);
 
-Eigen::Vector3f pqr2RPYDot(const Eigen::Ref<const Eigen::Vector3f> rpy, const Eigen::Ref<const Eigen::Vector3f> pqr);
+Eigen::Vector3d pqr2RPYDot(const Eigen::Ref<const Eigen::Vector3d> rpy, const Eigen::Ref<const Eigen::Vector3d> pqr);
 
-Eigen::Vector3f rpyDot2PQR(const Eigen::Ref<const Eigen::Vector3f> rpy, const Eigen::Ref<const Eigen::Vector3f> rpyDot);
+Eigen::Vector3d rpyDot2PQR(const Eigen::Ref<const Eigen::Vector3d> rpy, const Eigen::Ref<const Eigen::Vector3d> rpyDot);
 } // namespace rot3d
 } // namespace auv_core
